requirements.txt
Django>=2.2,<2.3
django-environ>=0.4.5,<0.5
psycopg2==2.8.3
# psycopg2-binary==2.8.3
djangorestframework==3.10.2
Pillow==6.1.0
coreapi==2.3.3

